Day 1, Solution 2
This modifies the data structure a bit to, instead of tracking a
`Player`'s final position, it tracks the position for every discrete
1-unit movement. With this, identifying the first time the line crosses
(since we're on a grid, it's always integer intersections) becomes
fairly trivial; given a list of all traveled positions, `firstRepeated`
will move through the list and find the first match. With that match, we
then calculate Manhattan distance.