Day 2, solution 1
I went into this with a sense that the keypad would be much larger for
round 2, so I wrote a parser to handle that, as well as the instructions
input.

Nothing interesting here, apart from the sequential aspect of handling
instructions - basically, that we need to track the final position given
a list of commands. Those are then aggregated into actual keypad presses
in the result.