Day 2, solution 2
This layers on additional keypad complexity by modifying the shape.

In this, I used a dash to represent a piece of the keypad without a
value, which resulted in the entire keypad being lists of `Maybe KeypadColumn`.
With the `Maybe` in place, the remaining piece to ensure staying within
bounds was updating `isValidPosition` to use all the previous boundary
checking, as well as checking that the space is a `Just KeypadColumn`
instead of `Nothing`.