Day 8, Solution 1 (and 2)
This uses Data.Array to build out the Board structure.

Parsing instructions was pretty straightforward, but nailing down the
operations for shifting rows/columns was pretty tricky. There's probably
a way to isolate runRotate better so there's some reuse, but I'm not
sure it's worth it.

Update: Because it prints the screen, this solves #2 as well